+++ src/Container/Providers/ConsoleProvider.php
@@ -34,6 +34,7 @@ final class ConsoleProvider extends BaseServiceProvider
             \Glueful\Console\Commands\Cache\TtlCommand::class,
             \Glueful\Console\Commands\Cache\ExpireCommand::class,
             \Glueful\Console\Commands\Cache\PurgeCommand::class,
+            \Glueful\Console\Commands\Cache\MaintenanceCommand::class,
             // Database commands
             \Glueful\Console\Commands\Database\StatusCommand::class,
             \Glueful\Console\Commands\Database\ResetCommand::class,